cercachain-main/ml/requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.30.0
ultralytics>=8.0.0  # For YOLOv9
facenet-pytorch>=2.5.0
spacy>=3.5.0

# Audio processing
openai-whisper>=20231117
deepspeech>=0.9.3
pydub>=0.25.1
numpy>=1.24.0

# Vision processing
opencv-python>=4.8.0
pytesseract>=0.3.10

# Utilities
tqdm>=4.65.0
Pillow>=10.0.0
loguru>=0.7.0

# Download spaCy model
# Run: python -m spacy download en_core_web_sm